Problem with DrawIndexed was that size_t was being used instead of int. Unsure of why this solves the problem.

+++ Salvage/Salvage/BoundingVolume.h
@@ -15,7 +15,7 @@ class BoundingVolume : public Drawable
 {
 private:
 	std::vector<BoundingVolumeVertex> _vertices;
-	std::vector<size_t> _indices;
+	std::vector<int> _indices;
 
 	ID3D11Buffer* _vertexBuffer;
 	ID3D11Buffer* _indexBuffer;
@@ -26,9 +26,9 @@ public:
 	BoundingVolume(ID3D11Device* device);
 	virtual ~BoundingVolume();
 	std::vector<BoundingVolumeVertex>* getVertices();
-	std::vector<size_t>* getIndices();
+	std::vector<int>* getIndices();
 	
-	void createBuffers(ID3D11Device* device, std::vector<BoundingVolumeVertex> vertices, std::vector<size_t> indices);
+	void createBuffers(ID3D11Device* device, std::vector<BoundingVolumeVertex> vertices, std::vector<int> indices);
 
 	void draw(ID3D11DeviceContext* deviceContext, ID3D11Buffer* transformationBuffer);
 };
